STC Viterbi memory optimization + encode/decode progress bars
32× memory reduction via 1-bit packed back pointers (u128 per step instead
of Vec<u32>), plus segmented checkpoint Viterbi for images >1M positions
(O(√n) memory, ~3 MB for any image size).

Enable real-time progress bars during encoding on iOS, Android, and WASM:
- Ghost encode: 12 steps (3 UNIWARD + 8 STC + 1 write)
- Armor encode: 6 steps (STDM) or 3 steps (Fortress)
- Rename compute_uniward_for_decode → compute_uniward_with_progress (shared)
- iOS/Android: remove !isEncode guard, poll same global atomics for both
- Add encode-specific status messages on all platforms
- Fix Try Decode passphrase field visibility (iOS + Android)
- Update README with Segmented Viterbi docs, progress API, memory tables

+++ src/stego/armor/dft_payload.rs
@@ -271,7 +271,7 @@ pub fn embed_ring_payload(
 
     // Encrypt payload
     let (ciphertext, nonce, salt) = crypto::encrypt(payload, passphrase)?;
-    let frame_bytes = frame::build_frame(payload.len() as u16, &salt, &nonce, &ciphertext);
+    let frame_bytes = frame::build_frame(payload.len(), &salt, &nonce, &ciphertext);
 
     // RS encode with heavy parity
     let rs_encoded = ecc::rs_encode_blocks_with_parity(&frame_bytes, RING_RS_PARITY);
